When an external command fails, say what the command was in the error
